Feature/create conan 1.0 package (#50)
* Integrate conan.io 1.0 into the cmake build files to simplify creation of conan.io packages
* Remove usage of Foonathans memory library due to missing conan.io integration and missing cross-compilation support
* Remove C++17 usage
* Allow to configure externals as parameter instead of using submodules
* Switch to main Abseil repository
* Switch to main memory library

+++ src/examples/meshviewer/components/gpusurfacemesh.cpp
@@ -48,11 +48,11 @@ void GPUSurfaceMesh::update()
 	// Create the index buffer
 	BufferDescription idxDesc;
 	idxDesc.Usage = ResourceUsage::Default;
-	idxDesc.SizeInBytes = tri_adjs.size() * _indexStride;
+	idxDesc.SizeInBytes = static_cast<uint32_t>(tri_adjs.size() * _indexStride);
 
 	BufferInitData idxData;
 	idxData.Data = tri_adjs.data();
-	idxData.SizeInBytes = tri_adjs.size() * _indexStride;
+	idxData.SizeInBytes = static_cast<uint32_t>(tri_adjs.size() * _indexStride);
 
 	_indices = std::make_unique<OpenGL::Buffer>(idxDesc, false, false, &idxData);
 
@@ -62,11 +62,11 @@ void GPUSurfaceMesh::update()
 	BufferDescription posDesc;
 	posDesc.CPUAccess = ResourceAccess::Write;
 	posDesc.Usage = ResourceUsage::Default;
-	posDesc.SizeInBytes = _triMesh->nrVertices() * _positionStride;
+	posDesc.SizeInBytes = static_cast<uint32_t>(_triMesh->nrVertices() * _positionStride);
 
 	BufferInitData posData;
 	posData.Data = _triMesh->vertices()->data();
-	posData.SizeInBytes = _triMesh->nrVertices() * _positionStride;
+	posData.SizeInBytes = static_cast<uint32_t>(_triMesh->nrVertices() * _positionStride);
 
 	_positions = std::make_unique<OpenGL::Buffer>(posDesc, false, false, &posData);
 
@@ -77,11 +77,11 @@ void GPUSurfaceMesh::update()
 	BufferDescription normalDesc;
 	normalDesc.CPUAccess = ResourceAccess::Write;
 	normalDesc.Usage = ResourceUsage::Default;
-	normalDesc.SizeInBytes = normals->size() * _normalStride;
+	normalDesc.SizeInBytes = static_cast<uint32_t>(normals->size() * _normalStride);
 
 	BufferInitData normalData;
 	normalData.Data = normals->data();
-	normalData.SizeInBytes = normals->size() * _normalStride;
+	normalData.SizeInBytes = static_cast<uint32_t>(normals->size() * _normalStride);
 
 	_normals = std::make_unique<OpenGL::Buffer>(normalDesc, false, false, &normalData);
